Update file Pathnames
Updated file pathnames so that the path says /Code/ and not /code/.  Noticing multiple files that are not present in dentalMeasures/dat or any other folder for that matter.  Either never got them from Jon or they are deprecated.  There is a file that involves Amanda measures that could be an older version of her data.