Update occFns.R
Updated getNOWDatesMatFromLocCodes with new code from Jon.  Updates the dates for Puercan 2 and 3 so they are now merged as per the ongoing work with NOW database (see email with Kari Lintulaakso and Indre Žliobaitė on 11/2/2023).  They are getting rid of the Puercan 3 so that the Pu/To boundary is earlier (64.81) as per Ogg and Gradstein 2016.