Update taxonomicEv.R
Worked on updating the fourClassesFromOccs function so that it works.  Made the origjnal version have a "_old" suffix and went to work making the function use current occs formats (e.g., occs$taxon was changed to occs$genus).  Still need to add way to parse the taxa list.  May need to use the restrict.col and restrict.class to do so (?....).